+++ .zed/settings.json
@@ -1,0 +1,20 @@
+// Folder-specific settings
+//
+// For a full list of overridable settings, and general information on folder-specific settings,
+// see the documentation: https://zed.dev/docs/configuring-zed#settings-files
+{
+  "languages": {
+    "TypeScript": {
+      "inlay_hints": {
+        "enabled": false
+      },
+      "language_servers": [
+        "biome",
+        "!tsgo",
+        "!eslint",
+        "!typescript-language-server",
+        "..."
+      ]
+    }
+  }
+}
