+++ bunup.config.ts
@@ -1,0 +1,32 @@
+import { type DefineConfigItem, defineWorkspace } from "bunup"
+import orchardPluginConfig from "./packages/orchard/bunup.config"
+import orchardCoreConfig from "./packages/orchard-core/bunup.config"
+import orchardRuntimeCfg from "./packages/orchard-runtime/bunup.config"
+
+const sharedOpts: Partial<DefineConfigItem> = {
+  clean: true,
+  dts: true,
+  format: "esm",
+  outDir: "dist",
+}
+
+export default defineWorkspace(
+  [
+    {
+      name: "@orchard/runtime",
+      root: "packages/orchard-runtime",
+      config: orchardRuntimeCfg,
+    },
+    {
+      name: "@orchard/core",
+      root: "packages/orchard-core",
+      config: orchardCoreConfig,
+    },
+    {
+      name: "@orchard/plugin",
+      root: "packages/orchard",
+      config: orchardPluginConfig,
+    },
+  ],
+  sharedOpts,
+)
