+++ internal/build-plugins/package.json
@@ -1,0 +1,22 @@
+{
+  "name": "@internal/build-plugins",
+  "description": "Build plugins for internal use",
+  "type": "module",
+  "private": true,
+  "exports": {
+    "./svelte": {
+      "bun": "./src/svelte.ts"
+    }
+  },
+  "devDependencies": {
+    "typescript": "catalog:",
+    "@typescript/native-preview": "catalog:",
+    "oxlint": "catalog:",
+    "oxlint-tsgolint": "catalog:",
+    "@biomejs/biome": "catalog:",
+    "@tsconfig/bun": "catalog:",
+    "bunup": "catalog:",
+    "svelte": "catalog:",
+    "@types/bun": "catalog:"
+  }
+}
