+++ packages/mcp-server/package.json
@@ -28,9 +28,9 @@
   "dependencies": {
     "@orchard/core": "workspace:*",
     "obsidian": "catalog:",
-    "hono": "^4.9.12",
+    "hono": "^4.10.1",
     "@hono/node-server": "^1.19.5",
-    "@modelcontextprotocol/sdk": "^1.20.0",
+    "@modelcontextprotocol/sdk": "^1.20.1",
     "zod": "^3.23.8"
   }
 }
