+++ packages/orchard-core/src/inline-task/index.ts
@@ -1,0 +1,12 @@
+export {
+  formatInlineTaskLine,
+  parseInlineTasks,
+  TASK_NOTE_TYPE,
+  type TaskSyncState,
+  type TaskFrontmatter,
+  type TaskFrontmatterInit,
+  type TaskCheckbox,
+  type InlineTaskExtraFields,
+  type InlineTask,
+  type InlineTaskFormatInput,
+} from "./formatter"
