+++ packages/orchard-runtime/biome.jsonc
@@ -1,5 +1,6 @@
 {
   "$schema": "../../node_modules/@biomejs/biome/configuration_schema.json",
+  "root": false,
   "files": {
     "includes": [
       "./jsonc",
